refactored code

+++ include/storage/disk_manager/disk_manager_async.hpp
@@ -44,7 +44,7 @@ namespace db7::storage
         bool OpenFile(table_id tbl_id);
         bool DropTable(table_id tbl_id);
         bool ExistsTable(table_id tbl_id);
-        bool TruncateFile(table_id tbl_id, u32 pages_num);
+        bool TruncateFile(table_id tbl_id, u64 pages_num);
         u32 PageCount(table_id tbl_id);
     };
 }
